modified guides

+++ apps/bookings/views.py
@@ -4,14 +4,21 @@ from django.shortcuts import render, get_object_or_404, redirect
 from .models import Booking
 from .forms import BookingForm
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, user_passes_test
 
-# @login_required
+
 def booking_list(request):
+    if "user_id" not in request.session:
+        return redirect("user_login:user-login")
+    
     bookings = Booking.objects.all()
     return render(request, 'bookings/booking_list.html', {'bookings': bookings})
 
-# @login_required
+
 def create_booking(request):
+    if "user_id" not in request.session:
+        return redirect("user_login:user-login")
+    
     if request.method == 'POST':
         form = BookingForm(request.POST)
         if form.is_valid():
@@ -21,8 +28,11 @@ def create_booking(request):
         form = BookingForm()
     return render(request, 'bookings/create_booking.html', {'form': form})
 
-# @login_required
+
 def edit_booking(request, booking_id):
+    if "user_id" not in request.session:
+        return redirect("user_login:user-login")
+    
     booking = get_object_or_404(Booking, booking_id=booking_id)
     if request.method == 'POST':
         form = BookingForm(request.POST, instance=booking)
@@ -33,8 +43,11 @@ def edit_booking(request, booking_id):
         form = BookingForm(instance=booking)
     return render(request, 'bookings/booking_form.html', {'form': form, 'booking': booking})
 
-@login_required
+
 def delete_booking(request, booking_id):
+    if "user_id" not in request.session:
+        return redirect("user_login:user-login")
+    
     booking = get_object_or_404(Booking, booking_id=booking_id)
     if request.method == 'POST':
         booking.delete()
